End Of Section 10

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Net;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,21 +8,27 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public required string Id { get; set; }
     }
 
-    public class Handler(AppDbContext context) : IRequestHandler<Command>
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
-            var activity = await context.Activities.FindAsync([request.Id], cancellationToken) 
-                ?? throw new Exception("Could not find activity");
+            var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
+            
+            if (activity == null)
+            {
+                return Result<Unit>.Failure("Activity not found", (int)HttpStatusCode.NotFound);
+            }
 
             context.Activities.Remove(activity);
+            var result = await context.SaveChangesAsync(cancellationToken) > 0;
 
-            await context.SaveChangesAsync(cancellationToken);
+            if (!result) return Result<Unit>.Failure("Fail to delete the activity", (int)HttpStatusCode.BadRequest);
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
